make the fd's used to communicate with childs non blocking
This fixes a grave bug where the daemon was unresponsive for many many
seconds. Especially during ext3 fs creation. Also read to end for
stderr and stdout when the child terminates. This fixes a long
standing bug where sometimes we didn't get all the data.

We now also pass stdout to the JobCompleted callback. Might come in
handy later.