tear down crypto devices when a device is unsafely removed
Streamline the whole way force removal works. It shouldn't be hard to
add force removal to RAID (e.g. md) and LVM arrays once we get around
to adding support for these.

We still need to check whether the encrypted device was set up by us
before tearing it down. That's for the next patch.