remove erase= option and the notion of tasks within jobs
Secure erase is a bit K-rad and we never really implemented anything
but 0 and 1 passes. If we want this feature back (and I think we do;
e.g. it is useful to scrub the disks before selling a computer) we
should add a dedicated SecureErase() method that (optionally) can take
advantage of things like the ATA Secure Erase command-set.

While it is cute that jobs consist of tasks it is not used anymore now
that erase= is gone. So remove it to simplify the code.