remove hack for working around broken USB card reader
This add was added since one of my card readers appear to be broken;
other devices are fine.

On closer investigation this looks like a kernel bug. Filed here for
now

http://bugzilla.kernel.org/show_bug.cgi?id=13029

If needed we can add code to do the equivalent of 'touch /dev/sdXXX'
in the poller on the first ENOMEDIUM for a device being polled. But
let's see what the kernel guys says before resorting to such hacks.