Bug 22149 – Do not poll floppy drives
Avoid polling PC floppy drives since that makes a lot of
noise. DrivePollMedia() still works, e.g.

 dbus-send --system --print-reply --dest=org.freedesktop.DeviceKit.Disks  \
           /org/freedesktop/DeviceKit/Disks/devices/fd0                   \
           org.freedesktop.DeviceKit.Disks.Device.DrivePollMedia

Note, though, that udev will not run /sbin/blkid on the device. So
desktops wanting to support this needs to ignore the fstype for floppy
drives and just attempt to mount it anyway.