Bug 24343 – Unintended side-effect when calling DriveDetach()
Some laptops like the Dell Mini 10 have builtin SD card readers which
are wired through an internal USB bus. The current behaviour is that
if you click the eject button in nautilus, the card reader device gets
ejected and disconnected, and cannot be reactivated ever, except by a
reboot.

The cause of this is that DK-d's update_info_drive() sets "can_detach"
for all USB devices currently, which causes
devkit-disks-helper-drive-detach to be called on them.

https://bugs.freedesktop.org/show_bug.cgi?id=24343