Handle loopback devices
 http://people.freedesktop.org/~david/udisks-loop-devices.png

it even works with kpartx partitions

 http://people.freedesktop.org/~david/palimpsest-loop-kpartx-partition.png

In order for this to work well, you will this patch

 http://people.freedesktop.org/~david/0001-Generate-change-uevent-for-loop-device.patch

which is also needed in order to make the /dev/disk hierarchy work
well. In lieu of this patch, you can generate the 'change' uevent
yourself. I've submitted this patch upstream through Kay Sievers.