Bug 26258 — initial btrfs support
Add basic support for creating and checking btrfs. This does not yet support
spanning multiple devices, RAID, etc.

We disable online fsck for now. It works, but is said to be too unstable still.

Also bump the size of the loop devices in the test suite, since btrfs needs at
least 256 MB.