testsuite: Move to scsi_debug
Drop the usage of loop devices and move to scsi_debug. This provides much more
realistic behaviour (such as partitioning) and avoids using a kind of device
which udisks special-cases. scsi_debug fake devices can also be forcibly
removed (unlike loop and md), thus we can write test cases to check the cleanup
in this case.

Since scsi_debug only provides one (independent) test device, reorganize the
code a little to either just use one that one device (for most tests such as FS
or Smart), or use two partitions on that (for the LVM PVs).

Drop the "Loop" test case; it is not that interesting.