Make parallel installable with udisks version 1
This includes renaming a bunch of stuff so we don't conflict. For now
we ignore collisions in man pages. We still have to rename object
paths and interfaces.

Also bump to version 1.90. First stable release will be 2.0.

Another huge user-visible change is that the udisks(1) command has
been renamed to udisksctl(1).

Signed-off-by: David Zeuthen <davidz@redhat.com>