Support calling device_drive_eject() and device_filesystem_unmount() from daemon
In device_drive_eject() and device_filesystem_unmount(), only specify a
PolicyKit action if we actually have a D-BUS invocation context. If not, the
method is called from within the daemon, so use a NULL PK action to avoid a
crash. This is similar to device_drive_ata_smart_refresh_data().

Use the throw_error() wrapper instead of direct dbus_g_method_return_error(),
to support NULL contexts.