Bug 34710 — CD-ROM polling failed due to O_EXCL flag
Newer kernels (2.6.38+) support in-kernel polling of CD-ROM/SD card devices.
udev 172 and later enable this feature by default:

  http://git.kernel.org/?p=linux/hotplug/udev.git;a=commitdiff;h=c5a41da949

That will also handle the eject button properly and send out remove uevents,
which causes stale mounts to be cleaned up properly.

In-kernel polling avoids the userspace race conditions with accessing the
drive.

If kernel polling is not supported or enabled, fall back to the previous method
of unlocking the CD drive door right after mounting, to keep the hardware
button working.