Change the default for LVM2 and dm-multipath support to off
The udev integration in device-mapper (and its users) is not exactly
working well - for example, you need to run 'udevadm trigger' in order
for the /lib/udev/udisks-lvm-pv-export prober to kick
in. Additionally, the fact that we run said prober on new events is
causing a lot of problems because its use of liblvm and the fact that
liblvm cannot currently be safely used in the udev path.

Signed-off-by: David Zeuthen <davidz@redhat.com>
Signed-off-by: Kay Sievers <kay@redhat.com>