Bug 41862 — Add label changing for previously supported file systems
Add support for label changing on vfat, ntfs, xfs, reiserfs, and nilfs2 file
systems.

Instead of having another huge if-then-else ladder as in udisks1, use a lookup
table _fs_info. For now this contains the commands for label changing and
erasing, but will be extended in the future to contain the commands for
creating and checking file systems.

https://bugs.freedesktop.org/show_bug.cgi?id=41862