For serial numbers, only use ID_SERIAL_SHORT, never ID_SERIAL
In the case that no real serial number exists, ID_SERIAL is made up
from vendor/model e.g. ST360021_A-0:0 or MediaTek_DVD-ROM_1309. We
don't want to convey such lies.

On the other hand, we should be guaranteed if ID_SERIAL_SHORT is set,
then it is something we can actually trust to be somewhat unique.

Signed-off-by: David Zeuthen <davidz@redhat.com>