Change some D-Bus types from bytestring ('ay') to UTF8 ('s')
We are 100% sure that the changed items are always UTF-8 especially
because udisks_device_udev_string() guarantees this and this function
is used before setting the affected properties.

This breaks the D-Bus API but there is no change in the libudisks2.so
ABI since byte-strings and UTF-8 strings are mapped the same way when
GDBus is being used. It may require restarting the system, but apart
from that there is minimal interference from this change.

The remaining byte-strings in the API are things that either comes
from configuration files (such as /etc/fstab and /etc/crypttab), the
kernel (e.g. sysfs paths) or udev rules (e.g. device symlinks). We
could require UTF-8 here and just reject items that are not valid
UTF-8 but there's really no point in that.

Signed-off-by: David Zeuthen <davidz@redhat.com>