Add multi-seat support
The first change is that we require additional authorization to
mount/unlock/modify a device, if it is on another seat than the
caller.

The other change is that we export the seat id as a property on the
org.freedesktop.UDisks2.Drive D-Bus interface. This can be used by
auto-mounter infrastructure (such as GVfs) to hint that a device
should not be automatically mounted unless it's plugged into the same
seat as where the environment is running.

All this is based on the libsystemd-login library - if this library is
not available at build-time, multi-seat support will not be enabled.

Signed-off-by: David Zeuthen <davidz@redhat.com>