Add loop.autoclear option to Filesystem.Mount() method
The main user of this could be the desktop shell / file manager
offering a "Attach Disk Image" option for e.g. ISO files - it would
first use Manager.LoopSetup() and then, on the returned block device,
invoke Filesystem.Mount() with this new option (assuming the returned
loop device has any interesting content). With this, the file manager
need not worry about the loop device - it goes away as soon as the
filesystem is unmounted.

Why is this an option to Filesystem.Mount() and not Manager.LoopSetup()?
Because right after setting up the loop device, the OS opens the
device so it can populate the udev database with the detected contents
of the newly created loop device and also to maintain symlinks in the
/dev/disk tree. If we had created the loop device with LO_AUTOCLEAR
the kernel would clear the loop device right after blkid(8) closes the
device. However, if we set the flag after successfully mounting the
file system on the loop device, there is already a single reference to
the block device.

Signed-off-by: David Zeuthen <davidz@redhat.com>