Bug 49842 – Unhandled rootfs on bind mount
Don't ignore a mount just because if a subtree of the filesystem is
mounted. Because if we do this, we may wrongly conclude the device is
not mounted at all, which then means that it's visible in the desktop
UI and subject to automounting.

In the future, we may want a "a(ss) MountPointsFull" property that in
addition to the mount point also contains the subtree that is mounted
at said mountpoint.

https://bugs.freedesktop.org/show_bug.cgi?id=49842

Signed-off-by: David Zeuthen <davidz@redhat.com>