Drive: Refuse to eject drives that appear to be in use
Unfortunately eject(1) is trying to be "helpful" insofar that it will
unmount partitions in use. While this is indeed handy when running
eject(1) from the command-line it's not a very good idea for us since
with the way its used in udisks, it effectively allows unmounting
partitions possibly mounted by other users.

So check if block devices for the drive is either mounted or unlocked
and bail if this is the case.

Signed-off-by: David Zeuthen <davidz@redhat.com>