Set Job:ExpectedEndTime when erasing a device
Do this by introducing some plumbing in UDisksBaseJob to estimate the
end time based on the UDisksJob:progress property. We do this by using
a 100-period moving average (actually, up to 100 periods). Works out
great in practice.

Also switch the format of the time used on the org.fd.UDisks.Job
interface to be micro-seconds instead of seconds. This is not really a
break as these properties were always zero in any released udisks2
version.

http://people.freedesktop.org/~david/gnome-disks-estimated-time-remaining.png

Signed-off-by: David Zeuthen <zeuthen@gmail.com>