integration-test: More robust syncing
Repeatedly settle the client until all main loop events have been processed,
instead of just calling it once. This avoids a lot of spurious, but obstinate
test failures due to properties not being updated yet.

Also add an extra sleep after the manual mount and dd call to allow some more
time for the changes to get picked up. This is a hack, but not easily fixed in
a cleaner way.