Add "Linux Filesystem" GPT partition type
According to http://en.wikipedia.org/wiki/GUID_Partition_Table there
is now a new dedicated Linux partition type so Microsoft's Basic Data
no longer has to be overloaded.

 Previously Linux used the same GUID for the data partitions as
 Windows. Linux never had a separate unique partition type GUID
 defined for its data partitions. This created problems when
 dual-booting Linux and Windows in UEFI-GPT setup. This new GUID was
 defined jointly by GPT fdisk and GNU Parted developers. It is
 identified as type code 8300 in GPT fdisk.

So add this new type and remove "Basic Data" from the Linux section
(it was listed in both the Linux and Microsoft section until now).

For now (and the foreseeable future) we will use "Basic Data" instead
of "Microsoft Basic Data" since a lot of systems in the wild is using
this type already and it would cause confusion to display the string
"Microsoft Basic Data" on Linux systems

Signed-off-by: David Zeuthen <zeuthen@gmail.com>