Ignore non-Linux software on SanDisk Cruzer
We've been happily ignoring this because of firmware-bugs in its
CD-ROM emulation but this may soon change, see

 https://bugs.freedesktop.org/show_bug.cgi?id=52474

In fact, newer revisions of the Cruzer (including the one Kay has) has
fixes for this (making the fake CD-ROM appear on the desktop).

The CD-ROM will still appear in GNOME Disks and it can be mounted that
way if you really want to access these files. It just won't appear in
the desktop, e.g. the file manager or file chooser until it's manually
mounted.

Signed-off-by: David Zeuthen <zeuthen@gmail.com>