integration-test: Split out and fix daemon startup
Split out start_daemon() and stop_daemon() into separate class methods, so that
tests can check properties that should persist across daemon restarts.

Also fix the startup to add --replace, to ensure that we are really talking to
the udisksd launched by the test suite, not to an already existing one running
in the system.