Introduce UDisksLinuxDevice as a replacement for GUdevDevice
This is essentially a GUdevDevice but with extra information such as
the probed ATA IDENTIFY data and, in the future, anything else we
might need.

The big idea here is that we want to depend _less_ on udev for
peripheral things that doesn't make sense to store in the udev
database, such as the properties in [1]. The win here is

 a. We can simplify ata_id somewhat - it still needs to extract serial
    numbers, WWNs etc. (because those are needed for symlinks in the
    /dev/disk/by-* hierarchy) but it does not need to extract and
    print what is in e.g. [1].

 b. If we need information in udisks from e.g. the ATA IDENTIFY data
    we don't need to add it to ata_id (or scsi_id or whatever), then
    wait for a systemd release and then instruct distributors to
    properly handle the dependency - instead, we can just access the
    IDENTIFY data ourselves. Much simpler and less error prone.

Since probing is synchronous, we create a thread for this and whenever
there's an uevent we push a new request onto a work queue. The probing
thread will handle incoming requests sequentially (in the future we
could have a pool of worker threads and run things in parallel) and
pop the results back to the main thread, also sequentially.

The only probing we have right now is for IDENTIFY DEVICE and IDENTIFY
PACKET DEVICE for ATA disks. We can add more in the future.

Note that this patch doesn't actually achieve b. yet - upcoming
patches will move from reading udev properties (such as looking at the
value of ID_ATA_FEATURE_SET_PM) to extracting this from the IDENTIFY
data itself.

[1] :
 # udevadm info -q all -n /dev/sda|grep ID_ATA_
 E: ID_ATA_DOWNLOAD_MICROCODE=1
 E: ID_ATA_FEATURE_SET_HPA=1
 E: ID_ATA_FEATURE_SET_HPA_ENABLED=1
 E: ID_ATA_FEATURE_SET_PM=1
 E: ID_ATA_FEATURE_SET_PM_ENABLED=1
 E: ID_ATA_FEATURE_SET_SECURITY=1
 E: ID_ATA_FEATURE_SET_SECURITY_ENABLED=0
 E: ID_ATA_FEATURE_SET_SECURITY_ENHANCED_ERASE_UNIT_MIN=2
 E: ID_ATA_FEATURE_SET_SECURITY_ERASE_UNIT_MIN=2
 E: ID_ATA_FEATURE_SET_SECURITY_FROZEN=1
 E: ID_ATA_FEATURE_SET_SMART=1
 E: ID_ATA_FEATURE_SET_SMART_ENABLED=1
 E: ID_ATA_ROTATION_RATE_RPM=0
 E: ID_ATA_SATA=1
 E: ID_ATA_SATA_SIGNAL_RATE_GEN1=1
 E: ID_ATA_SATA_SIGNAL_RATE_GEN2=1
 E: ID_ATA_WRITE_CACHE=1
 E: ID_ATA_WRITE_CACHE_ENABLED=1

Signed-off-by: David Zeuthen <zeuthen@gmail.com>