Block.Format: add @update-partition-type option
If this option is set to TRUE, we'll set the _natural_ partition type
for the requested filesystem, if one is available. For example, if
formatting a GPT partition with a FAT filesystem, the Microsoft Basic
Data partition type will be chosen; similar, if formatting a DOS
partition with a Ext4 filesystem then partition type 0x83 is chosen.

This feature was requested here

 https://bugzilla.gnome.org/show_bug.cgi?id=685871

Signed-off-by: David Zeuthen <zeuthen@gmail.com>