Don't bail in MD-RAID file monitor event handler
The code was set up to bail the first time the event fired which
wasn't what we wanted (doh!). So we didn't pick up any events at all.

Also change severity of the logging message to debug from warning,
since read errors happen sporadically when a MD device is removed.

Signed-off-by: David Zeuthen <zeuthen@gmail.com>