Send SCSI START STOP UNIT when powering down a drive
It turns out that some disks / enclosures will not spin down when the
power to the USB upstream port is removed. Therefore, issue the START
STOP UNIT command just before removing power. As this command does
fail on some devices [1] make the failure non-fatal. Also add some
logging - even on successful paths - as it's useful for the admin to
see when a drive has been powered down.

This was reported in bug 71802.

https://bugs.freedesktop.org/show_bug.cgi?id=71802

[1] : in fact, START STOP UNIT fails on one of my external USB drives
but works on another.

Signed-off-by: David Zeuthen <zeuthen@gmail.com>