Fix buffer overflow in mount path parsing
In the mount monitor we parse mount points from /proc/self/mountinfo and
/proc/swaps. Ensure that we don't overflow the buffers on platforms where mount
paths could be longer than PATH_MAX (unknown if that can actually happen), as
at least the mount paths for hotpluggable devices are somewhat user-controlled.

Thanks to Florian Weimer for discovering this bug, and to David Zeuthen
for his initial patch!

CVE-2014-0004