Fix standby timers
If the drive was set to standby after >= 10 minutes of inactivity,
it would never standby because udisks would check every 10 minutes
to see if it was already in standby, and if not, update the SMART
stats, which would reset the standby timer.  This is especially
problematic since many drives enforce a 10 minute minimum standby
timer even if a lower value is specified.

This patch checks the kernel IO stats to see if any other IO has
been done on the drive since the last check, and if not, behaves
as if the drive is already sleeping.

https://bugs.freedesktop.org/show_bug.cgi?id=80170
https://launchpad.net/bugs/1281588