Fix TOCTOU race when making directories
If two volumes are mounted at approximately the same time, one of the
mount operations can fail with an error something like:

  Error creating directory /run/media: File exists

because of the window between the check of the existence and the creation
of the directory. To fix this, drop the g_file_test() check and instead
ignore EEXIST errors when making these directories.

https://bugs.freedesktop.org/show_bug.cgi?id=83374