udisks-modular: Introduce plugin basics and block/device object interfaces
This commit introduces a simple plugin architecture. Modules are supposed
to be in-tree with expected usage of daemon header files. The main purpose
is just a delayed loading of extended functionality.

For simplicity, plugins are loaded and activated on daemon start. This will
change in the future though. All modules are supposed to be queried at the
same time, late activation should not be allowed. Also for the moment it's
supposed that modules will stay active for the rest of the daemon lifetime
to prevent issues on dlclose().

UDisksModuleManager is a singleton that carries the module handles and takes
care of loading.

As a first step the UDisksLinuxBlockObject and UDisksLinuxDriveObject classes
have been extended to process a list of interfaces supplied by modules that
are supposed to be exported on d-bus.