udisks-modular: Enable module loading on demand
This implements the originally wanted delayed module loading. The great
idea behind is to only activate modules when client is really able and
wanting to handle the extra interfaces coming from modules.

For compatibility and legacy clients no visible difference is made when
modules are not loaded. And since no extra information is expected, we
may actually save resources by not loading modules until first client
asks for it.

TODO: Offer daemon arguments like --disable-modules (for debugging) or
--force-load-modules (for easy development).