Fixed README.md
The "OVERVIEW" section in the README.md file points to the wrong
man page for information on API stability and intended audience of
Storaged. The sections "API STABILITY" and "AUDIENCE" can be found
in the storaged-project(8) man page (doc/man/storaged-project.xml)
and not in the storaged(8) man page (doc/man/storaged.xml). This
commit fixes that.