Block: Add "config-items" option to Format.
This can be used to make sure that fstab and crypttab is updated after
a format even when the client goes away.

Some config item fields are now optional to make it possible to
actually use the new "config-items" option, since proper values for
these elements are not known before the Format is complete.